[Tree widget]: Add sub-categories cache
