+++ packages/itwin/tree-widget/src/tree-widget-react/components/trees/classifications-tree/internal/ClassificationsTreeIdsCache.ts
@@ -7,6 +7,7 @@ import { defer, EMPTY, forkJoin, from, map, mergeMap, of, reduce, shareReplay, t
 import { Guid, Id64 } from "@itwin/core-bentley";
 import { ElementChildrenCache } from "../../common/internal/caches/ElementChildrenCache.js";
 import { ModelCategoryElementsCountCache } from "../../common/internal/caches/ModelCategoryElementsCountCache.js";
+import { SubCategoriesCache } from "../../common/internal/caches/SubCategoriesCache.js";
 import {
   CLASS_NAME_Classification,
   CLASS_NAME_ClassificationSystem,
@@ -16,7 +17,6 @@ import {
   CLASS_NAME_GeometricElement2d,
   CLASS_NAME_GeometricElement3d,
   CLASS_NAME_SpatialCategory,
-  CLASS_NAME_SubCategory,
 } from "../../common/internal/ClassNameDefinitions.js";
 import { catchBeSQLiteInterrupts } from "../../common/internal/UseErrorState.js";
 import { joinId64Arg } from "../../common/internal/Utils.js";
@@ -43,13 +43,13 @@ interface ClassificationInfo {
 /** @internal */
 export class ClassificationsTreeIdsCache implements Disposable {
   readonly #categoryElementCounts: ModelCategoryElementsCountCache;
-  #categorySubCategories: Observable<Map<CategoryId, Array<SubCategoryId>>> | undefined;
   #elementModelsCategories: Observable<Map<ModelId, { category2dIds: Id64Set; category3dIds: Id64Set; isSubModel: boolean }>> | undefined;
   #modelWithCategoryModeledElements: Observable<Map<ModelId, Map<CategoryId, Set<ElementId>>>> | undefined;
   #classificationInfos: Observable<Map<ClassificationId | ClassificationTableId, ClassificationInfo>> | undefined;
   #filteredElementsData: Observable<Map<ElementId, { modelId: Id64String; categoryId: Id64String; categoryOfTopMostParentElement: CategoryId }>> | undefined;
   #elementChildren2dCache: ElementChildrenCache;
   #elementChildren3dCache: ElementChildrenCache;
+  #subCategoriesCache: SubCategoriesCache;
   #queryExecutor: LimitingECSqlQueryExecutor;
   #hierarchyConfig: ClassificationsTreeHierarchyConfiguration;
   #componentId: GuidString;
@@ -75,6 +75,10 @@ export class ClassificationsTreeIdsCache implements Disposable {
       elementClassName: CLASS_NAME_GeometricElement3d,
       componentId: this.#componentId,
     });
+    this.#subCategoriesCache = new SubCategoriesCache({
+      queryExecutor: this.#queryExecutor,
+      componentId: this.#componentId,
+    });
   }
 
   public [Symbol.dispose]() {
@@ -89,47 +93,8 @@ export class ClassificationsTreeIdsCache implements Disposable {
     return (type === "2d" ? this.#elementChildren2dCache : this.#elementChildren3dCache).getAllChildElementsCount({ elementIds });
   }
 
-  private querySubCategories(): Observable<{ id: SubCategoryId; parentId: CategoryId }> {
-    return defer(() => {
-      const definitionsQuery = `
-        SELECT
-          sc.ECInstanceId id,
-          sc.Parent.Id categoryId
-        FROM
-          ${CLASS_NAME_SubCategory} sc
-        WHERE
-          NOT sc.IsPrivate
-      `;
-      return this.#queryExecutor.createQueryReader(
-        { ecsql: definitionsQuery },
-        { rowFormat: "ECSqlPropertyNames", limit: "unbounded", restartToken: `${this.#componentName}/${this.#componentId}/visible-sub-categories` },
-      );
-    }).pipe(
-      catchBeSQLiteInterrupts,
-      map((row) => {
-        return { id: row.id, parentId: row.categoryId };
-      }),
-    );
-  }
-  private getSubCategoriesInfo() {
-    this.#categorySubCategories ??= this.querySubCategories()
-      .pipe(
-        reduce((acc, queriedSubCategory) => {
-          const entry = acc.get(queriedSubCategory.parentId);
-          if (entry) {
-            entry.push(queriedSubCategory.id);
-          } else {
-            acc.set(queriedSubCategory.parentId, [queriedSubCategory.id]);
-          }
-          return acc;
-        }, new Map<CategoryId, Array<SubCategoryId>>()),
-      )
-      .pipe(shareReplay());
-    return this.#categorySubCategories;
-  }
-
   public getSubCategories(categoryId: Id64String): Observable<Array<SubCategoryId>> {
-    return this.getSubCategoriesInfo().pipe(map((categorySubCategories) => categorySubCategories.get(categoryId) ?? []));
+    return this.#subCategoriesCache.getSubCategories(categoryId);
   }
 
   private queryElementModelCategories(): Observable<{
